+++ packages/bloc_test/test/cubit_bloc_test_test.dart
@@ -18,14 +18,14 @@ void main() {
       blocTest<CounterCubit, int>(
         'emits [1] when increment is called',
         build: () => CounterCubit(),
-        act: (cubit) async => cubit.increment(),
+        act: (cubit) => cubit.increment(),
         expect: <int>[1],
       );
 
       blocTest<CounterCubit, int>(
         'emits [1] when increment is called with async act',
         build: () => CounterCubit(),
-        act: (cubit) async => cubit.increment(),
+        act: (cubit) => cubit.increment(),
         expect: <int>[1],
       );
 
@@ -33,7 +33,7 @@ void main() {
         'emits [1, 2] when increment is called multiple times'
         'with async act',
         build: () => CounterCubit(),
-        act: (cubit) async => cubit..increment()..increment(),
+        act: (cubit) => cubit..increment()..increment(),
         expect: <int>[1, 2],
       );
     });
@@ -48,7 +48,7 @@ void main() {
       blocTest<AsyncCounterCubit, int>(
         'emits [1] when increment is called',
         build: () => AsyncCounterCubit(),
-        act: (cubit) async => cubit.increment(),
+        act: (cubit) => cubit.increment(),
         expect: <int>[1],
       );
 
@@ -74,14 +74,14 @@ void main() {
       blocTest<DelayedCounterCubit, int>(
         'emits [] when increment is called without wait',
         build: () => DelayedCounterCubit(),
-        act: (cubit) async => cubit.increment(),
+        act: (cubit) => cubit.increment(),
         expect: <int>[],
       );
 
       blocTest<DelayedCounterCubit, int>(
         'emits [1] when increment is called with wait',
         build: () => DelayedCounterCubit(),
-        act: (cubit) async => cubit.increment(),
+        act: (cubit) => cubit.increment(),
         wait: const Duration(milliseconds: 300),
         expect: <int>[1],
       );
@@ -97,7 +97,7 @@ void main() {
       blocTest<InstantEmitCubit, int>(
         'emits [2] when increment is called',
         build: () => InstantEmitCubit(),
-        act: (cubit) async => cubit.increment(),
+        act: (cubit) => cubit.increment(),
         expect: <int>[2],
       );
 
@@ -105,7 +105,7 @@ void main() {
         'emits [2, 3] when increment is called'
         'multiple times with async act',
         build: () => InstantEmitCubit(),
-        act: (cubit) async => cubit..increment()..increment(),
+        act: (cubit) => cubit..increment()..increment(),
         expect: <int>[2, 3],
       );
     });
@@ -120,7 +120,7 @@ void main() {
       blocTest<MultiCounterCubit, int>(
         'emits [1, 2] when increment is called',
         build: () => MultiCounterCubit(),
-        act: (cubit) async => cubit.increment(),
+        act: (cubit) => cubit.increment(),
         expect: <int>[1, 2],
       );
 
@@ -128,7 +128,7 @@ void main() {
         'emits [1, 2, 3, 4] when increment is called'
         'multiple times with async act',
         build: () => MultiCounterCubit(),
-        act: (cubit) async => cubit..increment()..increment(),
+        act: (cubit) => cubit..increment()..increment(),
         expect: <int>[1, 2, 3, 4],
       );
     });
@@ -143,7 +143,7 @@ void main() {
       blocTest<ComplexCubit, ComplexState>(
         'emits [ComplexStateB] when emitB is called',
         build: () => ComplexCubit(),
-        act: (cubit) async => cubit.emitB(),
+        act: (cubit) => cubit.emitB(),
         expect: <Matcher>[isA<ComplexStateB>()],
       );
     });
@@ -164,7 +164,7 @@ void main() {
       blocTest<SideEffectCounterCubit, int>(
         'emits [1] when SideEffectCounterEvent.increment is called',
         build: () => SideEffectCounterCubit(repository),
-        act: (cubit) async => cubit.increment(),
+        act: (cubit) => cubit.increment(),
         expect: <int>[1],
         verify: (_) async {
           verify(repository.sideEffect()).called(1);
@@ -174,7 +174,7 @@ void main() {
       blocTest<SideEffectCounterCubit, int>(
         'does not require an expect',
         build: () => SideEffectCounterCubit(repository),
-        act: (cubit) async => cubit.increment(),
+        act: (cubit) => cubit.increment(),
         verify: (_) async {
           verify(repository.sideEffect()).called(1);
         },